requirements.txt
RPi.GPIO
opencv-python
numpy
imutils